+++ src/moltres/__init__.py
@@ -9,7 +9,7 @@ from .expressions import col, lit
 from .table.schema import column
 from .table.table import Database
 
-__version__ = "0.10.0"
+__version__ = "0.11.0"
 
 __all__ = [
     "AsyncDatabase",